functional MB1 now. Can run the cpu instr test rom. most seem to be correct

+++ src/cartridge.zig
@@ -34,5 +34,5 @@ pub const Cartridge = struct {
     romSize: RomSize,
     ramSize: RamSize,
     ramEnabled: bool,
-    bank: u8,
+    bank: usize,
 };
